+++ src/main/java/com/tempomate/interceptor/LoginCheckInterceptor.java
@@ -7,7 +7,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -17,8 +16,8 @@ import javax.servlet.http.HttpServletResponse;
 public class LoginCheckInterceptor implements HandlerInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest req, HttpServletResponse resp, Object handler) throws Exception {
-        System.out.println("进入拦截器");
-        //获得请求头里的token
+        log.info("Enter interceptor");
+        //获得请求头里的token Get the token in the request header
         String jwt = req.getHeader("token");
 
         //判断token是否为空 Determine whether the token is empty